singular/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
anthropic==0.28.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
